[Refactor] Rearrange cafe directory structure
